refactor: 폴더 구조 수정(도메인 로직끼리 분리될 수 있는 방법 채택)
