+++ src/domain/savings-products/business/calculation.ts
