+++ src/domain/savings-products/components/SavingsProductPanel.tsx
