+++ src/domain/savings-products/queries.ts
@@ -1,6 +1,6 @@
 import { useSuspenseQuery } from '@tanstack/react-query';
 import { QUERY_KEY } from 'contants/queryKey';
-import { fetchSavingsProducts } from './http';
+import { fetchSavingsProducts } from './api';
 
 export const useFetchSavingsProducts = () => {
   return useSuspenseQuery({
